+++ src/dsdl/mixer/soundbank.d
@@ -24,7 +24,7 @@ class SoundBank : ResourceBank!SoundChunk {
             return;
         }
         foreach (ref string line; lns) {
-            uint indexOfSpace = indexOfAny(line, " \t");
+            long indexOfSpace = indexOfAny(line, " \t");
             if (indexOfSpace == -1 || indexOfSpace == line.length)
                 throw new DataFormatException("Unable to parse file \"" ~ chunkListFile ~ "\"");
             string chunkName = line[0..indexOfSpace];
@@ -36,4 +36,3 @@ class SoundBank : ResourceBank!SoundChunk {
     }
 
 }
-
